Don't update public directory during `app:update` command for API-only Applications
We remove files under the public directory for API-only Applications.
https://github.com/rails/rails/blob/a17aa6376d7974e38af90a2fe28e548b201baa63/railties/lib/rails/generators/rails/app/app_generator.rb#L500-L509

We should keep the behavior in `app:update` command too.

Follow up of #51952.
