+++ railties/lib/rails/generators/rails/app/app_generator.rb
@@ -224,6 +224,8 @@ module Rails
     end
 
     def public_directory
+      return if options[:update] && options[:api]
+
       directory "public", "public", recursive: false
     end
